+++ pkg/driver/driver.go
@@ -1,0 +1,3 @@
+package driver
+
+type Driver interface{}
